新規登録のページをコンポーネント化
